refactored the docs building scripts
